TestDocker/reflection/requirements.txt
numpy
pyrealsense2
torch
torchvision
opencv-python
pycocotools
python-socketio==4.4.0
requests
